# api/openapi.yaml
openapi: 3.0.3
info:
  title: POS System API
  description: API for Point of Sale System with user management, product management, and sales reporting
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: uint
        image:
          type: string
        name:
          type: string
        sku:
          type: string
        harga_modal:
          type: number
          format: float64
        harga_jual:
          type: number
          format: float64
        stock:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    UpdateProductRequest:
      type: object
      properties:
        image:
          type: string
        name:
          type: string
        sku:
          type: string
        harga_modal:
          type: number
          format: float64
        harga_jual:
          type: number
          format: float64
        stock:
          type: integer
    UpdateStockRequest:
      type: object
      required:
        - stock
      properties:
        stock:
          type: integer
          minimum: 0
...
paths:
...
  /products:
    get:
      summary: List all products
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint
    get:
      summary: Get a product by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      summary: Update a product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{id}/stock:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: uint
    put:
      summary: Update product stock
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
      responses:
        '200':
          description: Product stock updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
...